Refactor sstudio types.
