+++ smarts/sstudio/genscenario.py
@@ -28,7 +28,7 @@ import logging
 import os
 import pickle
 import sqlite3
-from dataclasses import asdict, replace
+from dataclasses import asdict, dataclass, replace
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
@@ -44,6 +44,18 @@ from .generators import TrafficGenerator
 logging.basicConfig(level=logging.WARNING)
 logger = logging.getLogger(__file__)
 logger.setLevel(logging.WARNING)
+
+
+@dataclass(frozen=True)
+class ActorAndMission:
+    """Holds an Actor object and its associated Mission."""
+
+    actor: types.Actor
+    """Specification for traffic actor.
+    """
+    mission: Union[types.Mission, types.EndlessMission, types.LapMission]
+    """Mission for traffic actor.
+    """
 
 
 def _check_if_called_externally():
@@ -592,7 +604,7 @@ def gen_missions(
     _validate_missions(missions)
 
     missions = [
-        types.ActorAndMission(actor=actor, mission=resolve_mission(mission))
+        ActorAndMission(actor=actor, mission=resolve_mission(mission))
         for actor, mission in itertools.product(actors, missions)
     ]
     with open(output_path, "wb") as f:
