+++ smarts/sstudio/tests/baseline.rou.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!-- generated on 2023-04-14 23:17:43 by Eclipse SUMO duarouter Version 1.16.0
+<!-- generated on 2023-05-15 16:41:19 by Eclipse SUMO duarouter Version 1.17.0
 <configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/duarouterConfiguration.xsd">
 
     <input>
         <net-file value="scenarios/sumo/intersections/4lane_t/map.net.xml"/>
-        <route-files value="/tmp/tmpiy_i5cen/trips.trips.xml"/>
+        <route-files value="/tmp/tmpppjoy0ji/trips.trips.xml"/>
     </input>
 
     <output>
-        <output-file value="smarts/sstudio/tests/traffic/generated.rou.xml"/>
+        <output-file value="/tmp/tmpc0mshfsm/traffic/generated.rou.xml"/>
     </output>
 
     <processing>
@@ -18,7 +18,8 @@
     </processing>
 
     <report>
-        <error-log value="/home/dev/.smarts/_duarouter_routing/traffic/generated.log"/>
+        <error-log value="/home/mtuck/.smarts/_duarouter_routing/traffic/generated.log"/>
+        <ignore-errors value="false"/>
         <no-step-log value="true"/>
     </report>
 
@@ -30,18 +31,18 @@
 -->
 
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/routes_file.xsd">
-    <vType id="car--61546" minGap="2.50" maxSpeed="55.50" speedFactor="normc(1.00,0.20,0.20,2.00)" vClass="passenger" accel="2.6" decel="4.5" sigma="0.5"/>
-    <vehicle id="car-edge-east-EW_0_30-edge-west-EW_0_-30-454981-765358-1-0.0" type="car--61546" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
+    <vType id="car-472435" minGap="2.50" maxSpeed="55.50" speedFactor="normc(1.00,0.20,0.20,2.00)" vClass="passenger" accel="2.6" decel="4.5" sigma="0.5"/>
+    <vehicle id="car-edge-east-EW_0_30-edge-west-EW_0_-30-237307--14548-1-0.0" type="car-472435" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
         <route edges="edge-east-EW edge-west-EW"/>
     </vehicle>
-    <vType id="car--37649" minGap="2.50" maxSpeed="55.50" speedFactor="normc(0.80,0.20,0.20,2.00)" vClass="passenger" impatience="0.50" lcCooperative="0.25" lcImpatience="1" jmDriveAfterYellowTime="1.0" accel="2.6" decel="4.5" sigma="0.5"/>
-    <vehicle id="car-edge-east-EW_0_30-edge-west-EW_0_-30-454981-765358-1-1.0" type="car--37649" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
+    <vType id="car-819925" minGap="2.50" maxSpeed="55.50" speedFactor="normc(0.80,0.20,0.20,2.00)" vClass="passenger" impatience="0.50" lcCooperative="0.25" lcImpatience="1" jmDriveAfterYellowTime="1.0" accel="2.6" decel="4.5" sigma="0.5"/>
+    <vehicle id="car-edge-east-EW_0_30-edge-west-EW_0_-30-237307--14548-1-1.0" type="car-819925" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
         <route edges="edge-east-EW edge-west-EW"/>
     </vehicle>
-    <vehicle id="car-edge-west-WE_0_30-edge-east-WE_0_-30-822864-585157-0-0.0" type="car--61546" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
+    <vehicle id="car-edge-west-WE_0_30-edge-east-WE_0_-30--66600-534261-0-0.0" type="car-472435" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
         <route edges="edge-west-WE edge-east-WE"/>
     </vehicle>
-    <vehicle id="car-edge-west-WE_0_30-edge-east-WE_0_-30-822864-585157-0-1.0" type="car--37649" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
+    <vehicle id="car-edge-west-WE_0_30-edge-east-WE_0_-30--66600-534261-0-1.0" type="car-819925" depart="0.00" departLane="0" departPos="30.00" departSpeed="max" arrivalLane="0" arrivalPos="-30.00">
         <route edges="edge-west-WE edge-east-WE"/>
     </vehicle>
 </routes>
