+++ smarts/sstudio/types/__init__.py
@@ -1,0 +1,35 @@
+# Copyright (C) 2020. Huawei Technologies Co., Ltd. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+from smarts.sstudio.types.actor.social_agent_actor import *
+from smarts.sstudio.types.actor.traffic_actor import *
+from smarts.sstudio.types.bubble import *
+from smarts.sstudio.types.bubble_limits import *
+from smarts.sstudio.types.condition import *
+from smarts.sstudio.types.constants import *
+from smarts.sstudio.types.dataset import *
+from smarts.sstudio.types.distribution import *
+from smarts.sstudio.types.entry_tactic import *
+from smarts.sstudio.types.map_spec import *
+from smarts.sstudio.types.mission import *
+from smarts.sstudio.types.route import *
+from smarts.sstudio.types.scenario import *
+from smarts.sstudio.types.traffic import *
+from smarts.sstudio.types.traffic_model import *
+from smarts.sstudio.types.zone import *
